requirements.txt
fastapi==0.94.1
uvicorn==0.21.1
pydantic==1.10.6
psycopg2-binary==2.9.5